backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==4.1.0
colorama==0.4.6
comm==0.2.3
comtypes==1.4.11
contourpy==1.3.3
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
fastapi==0.119.0
fastjsonschema==2.21.2
filelock==3.16.1
fire==0.7.0
firebase-admin==6.9.0
Flask==3.0.3
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
hyperframe==6.1.0
idna==3.7
imageio==2.37.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.1
openai==0.28.0
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pdf2docx==0.5.8
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pygame==2.6.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
PyMuPDF==1.26.1
pypandoc==1.15
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytube==15.0.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.5.1
scikit-fuzzy==0.5.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.0
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.38.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.14
xlsxwriter==3.2.5
yarl==1.18.3
yt-dlp==2025.8.22
